Move MockMvcUtils to mocks package
