+++ service/src/test/java/bio/terra/workspace/common/mocks/MockGitRepoApi.java
@@ -1,7 +1,6 @@
 package bio.terra.workspace.common.mocks;
 
 import bio.terra.workspace.common.fixtures.ReferenceResourceFixtures;
-import bio.terra.workspace.common.utils.MockMvcUtils;
 import bio.terra.workspace.generated.model.ApiCloneReferencedGitRepoResourceResult;
 import bio.terra.workspace.generated.model.ApiCloningInstructionsEnum;
 import bio.terra.workspace.generated.model.ApiCreateGitRepoReferenceRequestBody;
