+++ service/src/test/java/bio/terra/workspace/service/danglingresource/DanglingResourceCleanupServiceTest.java
@@ -9,8 +9,8 @@ import bio.terra.cloudres.google.dataproc.DataprocCow;
 import bio.terra.workspace.app.configuration.external.DanglingResourceCleanupConfiguration;
 import bio.terra.workspace.common.BaseConnectedTest;
 import bio.terra.workspace.common.mocks.MockGcpApi;
+import bio.terra.workspace.common.mocks.MockMvcUtils;
 import bio.terra.workspace.common.utils.GcpUtils;
-import bio.terra.workspace.common.utils.MockMvcUtils;
 import bio.terra.workspace.common.utils.TestUtils;
 import bio.terra.workspace.connected.UserAccessUtils;
 import bio.terra.workspace.connected.WorkspaceConnectedTestUtils;
