+++ service/src/test/java/bio/terra/workspace/service/resource/controlled/cloud/aws/sageMakerNotebook/CreateDeleteAwsSageMakerNotebookStepTest.java
@@ -24,8 +24,8 @@ import bio.terra.stairway.StepStatus;
 import bio.terra.workspace.common.exception.InternalLogicException;
 import bio.terra.workspace.common.fixtures.ControlledAwsResourceFixtures;
 import bio.terra.workspace.common.fixtures.WorkspaceFixtures;
+import bio.terra.workspace.common.mocks.MockMvcUtils;
 import bio.terra.workspace.common.utils.AwsUtils;
-import bio.terra.workspace.common.utils.MockMvcUtils;
 import bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys.ControlledResourceKeys;
 import org.junit.jupiter.api.Test;
 import org.mockito.MockedStatic;
